Century From Year.

+++ arcade-mode/102- centuryFromYear/solution.php
@@ -1,15 +1,11 @@
 <?php
 
-/**
- * @param int $year
- * @return int
- */
 function centuryFromYear($year)
 {
     $end = substr($year, -2, 2);
     if ($end == '00') {
-        return floor($year / 100);
+        return (int) floor($year / 100);
     } else {
-        return floor($year / 100) + 1;
+        return (int) floor($year / 100) + 1;
     }
 }
